+++ src/Crypto/RNG/Utils.hs
@@ -6,8 +6,16 @@ import Data.Primitive.SmallArray
 import Crypto.RNG
 
 -- | Generate random string of specified length that contains allowed chars.
+--
+-- The list of allowed chars must not be empty.
 randomString :: CryptoRNG m => Int -> [Char] -> m String
-randomString n allowedList = map (indexSmallArray allowed)
-  <$> replicateM n (randomR (0, sizeofSmallArray allowed - 1))
+randomString n allowedList
+  | size == 0 = error "randomString: the list of allowed chars is empty"
+  | otherwise = map (indexSmallArray allowed)
+      <$> replicateM n (randomR (0, size - 1))
   where
+    allowed :: SmallArray Char
     allowed = smallArrayFromList allowedList
+
+    size :: Int
+    size = sizeofSmallArray allowed